requirements.txt
django
djangorestframework
requests
scapy